Feature: Allow the SNMP Agent to be disabled
This feature is for larger sites who do not wish to run the SNMP Agent
in cacti.

+++ lib/snmpagent.php
@@ -22,8 +22,16 @@
    +-------------------------------------------------------------------------+
 */
 
+function snmpagent_enabled() {
+	return read_config_option('enable_snmp_agent') == 'on';
+}
+
 function snmpagent_cacti_stats_update($data){
 	$mc = new MibCache();
+
+	if (!snmpagent_enabled()) {
+		return false;
+	}
 
 	/* refresh total stats */
 	$mc->object('cactiStatsTotalsDevices')->set( snmpagent_read('cactiStatsTotalsDevices') );
@@ -88,6 +96,10 @@ function snmpagent_global_settings_update(){
 }
 
 function snmpagent_api_device_new($device){
+	if (!snmpagent_enabled()) {
+		return false;
+	}
+
 	$mc = new MibCache();
 	/* add device to cactiApplDeviceTable and cactiStatsDeviceTable*/
 	$device_data = db_fetch_row_prepared('SELECT * FROM `host` WHERE id = ?', array($device['id']));
@@ -138,6 +150,10 @@ function snmpagent_api_device_new($device){
 
 
 function snmpagent_data_source_action_bottom($data){
+	if (!snmpagent_enabled()) {
+		return false;
+	}
+
 	$mc = new MibCache();
 	$action = $data[0];
 	if ($action == '1') {
@@ -153,6 +169,10 @@ function snmpagent_data_source_action_bottom($data){
 }
 
 function snmpagent_graphs_action_bottom($data){
+	if (!snmpagent_enabled()) {
+		return false;
+	}
+
 	$mc = new MibCache();
 	$action = $data[0];
 	if ($action == '1') {
@@ -168,6 +188,10 @@ function snmpagent_graphs_action_bottom($data){
 }
 
 function snmpagent_device_action_bottom($data){
+	if (!snmpagent_enabled()) {
+		return false;
+	}
+
 	$mc = new MibCache();
 	$action = $data[0];
 	$selected_items = $data[1];
@@ -263,6 +287,10 @@ function snmpagent_device_action_bottom($data){
 }
 
 function snmpagent_poller_exiting($poller_index = 1){
+	if (!snmpagent_enabled()) {
+		return false;
+	}
+
 	$mc = new MibCache();
 
 	try
@@ -285,6 +313,14 @@ function snmpagent_poller_exiting($poller_index = 1){
 
 function snmpagent_poller_bottom() {
 	global $config;
+
+	if (!snmpagent_enabled()) {
+		snmpagent_cache_unistall();
+
+		return false;
+	} elseif (!snmpagent_cache_initialized()) {
+		snmpagent_cache_rebuilt();
+	}
 
 	if (api_plugin_is_enabled('maint')) {
 		include_once($config['base_path'] . '/plugins/maint/functions.php');
@@ -560,12 +596,16 @@ function snmpagent_get_pluginslist(){
 }
 
 /**
- * snmpagent_cache_setup()
+ * snmpagent_cache_install()
  * Generates a SNMP caching tables reflecting all objects of the Cacti MIB
  * @return
  */
-function snmpagent_cache_install(){
+function snmpagent_cache_install() {
 	global $config;
+
+	if (!snmpagent_enabled()) {
+		return false;
+	}
 
 	/* drop everything */
 	db_execute('TRUNCATE `snmpagent_cache`');
@@ -581,6 +621,18 @@ function snmpagent_cache_install(){
 
 	/* call install routine of plugins supporting the SNMPagent */
 	api_plugin_hook('snmpagent_cache_install');
+}
+
+function snmpagent_cache_uninstall() {
+	/* drop everything */
+	db_execute('TRUNCATE `snmpagent_cache`');
+	db_execute('TRUNCATE `snmpagent_mibs`;');
+	db_execute('TRUNCATE `snmpagent_cache_notifications`;');
+	db_execute('TRUNCATE `snmpagent_cache_textual_conventions`;');
+}
+
+function snmpagent_cache_initialized() {
+	return db_fetch_cell('SELECT COUNT(*) FROM `snmpagent_cache`') > 0;
 }
 
 function snmpagent_cache_rebuilt(){
@@ -606,7 +658,7 @@ function snmpagent_cache_init(){
 		foreach($pollers as $poller){
 			$poller_data = db_fetch_row_prepared('SELECT * FROM poller WHERE id = ?', array($poller['id']));
 		}
-	}else {
+	} else {
 		/* this is NOT a distributed system, but it should have at least one local poller. */
 		$poller_lastrun = read_config_option('poller_lastrun', true);
 		$values = array(
@@ -767,7 +819,7 @@ function snmpagent_notification($notification, $mib, $varbinds, $severity = SNMP
 		cacti_log('ERROR: Unknown event: ' . $notification . ' (' . $mib . ')', false, 'SNMPAGENT', POLLER_VERBOSITY_NONE);
 
 		return false;
-	}else {
+	} else {
 		$branches = explode('.', $enterprise_oid);
 		$specific_trap_number = array_pop($branches);
 	}
@@ -786,7 +838,7 @@ function snmpagent_notification($notification, $mib, $varbinds, $severity = SNMP
 		/* To bad! Nobody wants to hear our message. :( */
 		if (in_array($severity, array(SNMPAGENT_EVENT_SEVERITY_HIGH, SNMPAGENT_EVENT_SEVERITY_CRITICAL))) {
 			cacti_log('WARNING: No notification receivers configured for event: ' . $notification . ' (' . $mib . ')', false, 'SNMPAGENT', POLLER_VERBOSITY_NONE);
-		}else {
+		} else {
 			/* keep notifications of a lower/medium severity in mind to make a quicker decision next time */
 			$config['snmpagent']['notifications']['ignore'][$notification] = 1;
 		}
@@ -900,7 +952,7 @@ function snmpagent_notification($notification, $mib, $varbinds, $severity = SNMP
 						$snmp_security_level = 'authPriv';
 					} elseif ( $notification_manager['snmp_password'] && !$notification_manager['snmp_priv_passphrase']) {
 						$snmp_security_level = 'authNoPriv';
-					}else {
+					} else {
 						$snmp_security_level = 'noAuthNoPriv';
 					}
 					$args .= ' -l ' . $snmp_security_level . (($snmp_security_level != 'noAuthNoPriv') ? ' -a ' . $notification_manager['snmp_auth_protocol'] . ' -A ' . $notification_manager['snmp_password'] : '' ) . (($snmp_security_level == 'authPriv')? ' -x ' . $notification_manager['snmp_priv_protocol'] . ' -X ' . $notification_manager['snmp_priv_passphrase'] : '')  . ' ' . $notification_manager['hostname'] . ':' . $notification_manager['snmp_port'] . " \"\" " . $enterprise_oid . $snmp_notification_varbinds;
@@ -925,7 +977,7 @@ function snmpagent_notification($notification, $mib, $varbinds, $severity = SNMP
 				cacti_log("NOTE: $path_snmptrap " . str_replace(array($notification_manager['snmp_password'], $notification_manager['snmp_priv_passphrase']), '********', $args), false, 'SNMPAGENT', POLLER_VERBOSITY_MEDIUM);
 			}
 		}
-	}else {
+	} else {
 		/* mismatching number of var binds */
 		cacti_log('ERROR: Incomplete number of varbinds given for event: ' . $notification . ' (' . $mib . ')', false, 'SNMPAGENT', POLLER_VERBOSITY_NONE);
 		return false;
